Moved all inbox .NET implementations to a corresponding OwlCore.Storage.System.* namespace
See https://github.com/Arlodotexe/OwlCore.Storage/issues/48
