+++ tests/OwlCore.Storage.Tests/StreamFileTests.cs
@@ -1,4 +1,5 @@
 using OwlCore.Storage.CommonTests;
+using OwlCore.Storage.System.IO;
 
 namespace OwlCore.Storage.Tests
 {
